feat(calendar): Correctly splitting highlighted range
